🐛 fix script to get correct filepath
If script was run outside of repo it would fail because normally shell
uses $PWD and executes cmds from there, the chezmoi.yaml.tmpl could be
only initialized if `chezmoi -S custompath init` was run within the
repo. But not outside. This addition will fetch the filepath of the
script itself and use that for the git command.