✨ add usbipclient script
* add attaching and detaching of usb devices you want that are shareable
  from an usbipd server
* runs via systemctl service usbip.service

Issue:
When starting up:
Works so that you need to first run the server to startup then run the
client to attach the device.

Works shutting down:
Works so that you first need to shutdown the client and then the server.

If done differently usbip might end up in a bugged state, where you need
to manually interventions.

TODO:
* Test out what happens if order of boot/shutdown is done wrong and how
  to fix that.
* Try to think of a solution to make this a smoother experience, maybe
  with pinging the server and client and if they response keep going.
  But if not run the manual intervention script (Which I would need to
  create).